+++ passivbot/config.yaml
@@ -1,0 +1,48 @@
+#
+# copy this to manager/config.yaml and edit
+# structure:
+#
+defaults:
+  # name of a file in the config/live directory
+  live_config_name: "config.json"
+
+  # absolute path. Will be auto-generated if empty
+  live_config_path: ""
+  market_type: "futures" # futures | spot
+
+  # 0 means no limit
+  long_wallet_exposure_limit: 0
+  short_wallet_exposure_limit: 0
+  assigned_balance: 0
+
+  # modes:  # n | m | gs | p | t
+  # modes info: https://www.passivbot.com/en/latest/live/#controlling-the-bot
+  long_mode: "gs"
+  short_mode: "gs"
+
+instances:
+  - user: "binance_01"
+    symbols:
+      - "1000LUNCBUSD"
+      - "ADABUSD"
+      - "ANCBUSD"
+      - "APEBUSD"
+      - "AVAXBUSD"
+      - "BNBBUSD"
+      - "DODOBUSD"
+      - "DOGEBUSD"
+      - "DOTBUSD"
+      - "ETHBUSD"
+      - "FTMBUSD"
+      - "FTTBUSD"
+      - "GALABUSD"
+      - "GALBUSD"
+      - "GMTBUSD"
+      - "ICPBUSD"
+      - "LINKBUSD"
+      - "LUNA2BUSD"
+      - "NEARBUSD"
+      - "TLMBUSD"
+      - "TRXBUSD"
+      - "WAVESBUSD"
+      - "XRPBUSD"
