+++ session.yaml
@@ -6,126 +6,16 @@ windows:
       - cd /root/apollo
     panes:
       - shell_command:
-          - python3 random/1000LUNCBUSD.py
+          - python3 random.py
       - shell_command:
-          - sleep 150s
-          - python3 random/ADABUSD.py
-      - shell_command:
-          - sleep 300s
-          - python3 random/ANCBUSD.py
-      - shell_command:
-          - sleep 450s
-          - python3 random/APEBUSD.py
-      - shell_command:
-          - sleep 600s
-          - python3 random/AVAXBUSD.py
-      - shell_command:
-          - sleep 750s
-          - python3 random/BNBBUSD.py
-      - shell_command:
-          - sleep 900s
-          - python3 random/DODOBUSD.py
-      - shell_command:
-          - sleep 1050s
-          - python3 random/DOGEBUSD.py
-      - shell_command:
-          - sleep 1200s
-          - python3 random/DOTBUSD.py
-      - shell_command:
-          - sleep 1350s
-          - python3 random/ETHBUSD.py
-      - shell_command:
-          - sleep 1500s
-          - python3 random/FTMBUSD.py
-      - shell_command:
-          - sleep 1650s
-          - python3 random/FTTBUSD.py
-      - shell_command:
-          - sleep 1800s
-          - python3 random/GALABUSD.py
-      - shell_command:
-          - sleep 1950s
-          - python3 random/GALBUSD.py
-      - shell_command:
-          - sleep 2100s
-          - python3 random/GMTBUSD.py
-      - shell_command:
-          - sleep 2250s
-          - python3 random/ICPBUSD.py
-      - shell_command:
-          - sleep 2400s
-          - python3 random/LINKBUSD.py
-      - shell_command:
-          - sleep 2550s
-          - python3 random/LUNA2BUSD.py
-      - shell_command:
-          - sleep 2700s
-          - python3 random/NEARBUSD.py
-      - shell_command:
-          - sleep 2850s
-          - python3 random/TLMBUSD.py
-      - shell_command:
-          - sleep 3000s
-          - python3 random/TRXBUSD.py
-      - shell_command:
-          - sleep 3150s
-          - python3 random/WAVESBUSD.py
-      - shell_command:
-          - sleep 3300s
-          - python3 random/XRPBUSD.py
-  - window_name: editor_and_gs
+          - python3 variables_editor.py
+  - window_name: gs
     layout: even-vertical
     shell_command_before:
       - cd /root/passivbot
     panes:
       - shell_command:
-          - python3 /root/apollo/variables_editor.py
-      - shell_command:
-          - python3 passivbot.py binance_01 1000LUNCBUSD /root/apollo/config.json -lm gs -sm gs
-      - shell_command:
-          - python3 passivbot.py binance_01 ADABUSD /root/apollo/config.json -lm gs -sm gs
-      - shell_command:
-          - python3 passivbot.py binance_01 ANCBUSD /root/apollo/config.json -lm gs -sm gs
-      - shell_command:
-          - python3 passivbot.py binance_01 APEBUSD /root/apollo/config.json -lm gs -sm gs
-      - shell_command:
-          - python3 passivbot.py binance_01 AVAXBUSD /root/apollo/config.json -lm gs -sm gs
-      - shell_command:
-          - python3 passivbot.py binance_01 BNBBUSD /root/apollo/config.json -lm gs -sm gs
-      - shell_command:
-          - python3 passivbot.py binance_01 DODOBUSD /root/apollo/config.json -lm gs -sm gs
-      - shell_command:
-          - python3 passivbot.py binance_01 DOGEBUSD /root/apollo/config.json -lm gs -sm gs
-      - shell_command:
-          - python3 passivbot.py binance_01 DOTBUSD /root/apollo/config.json -lm gs -sm gs
-      - shell_command:
-          - python3 passivbot.py binance_01 ETHBUSD /root/apollo/config.json -lm gs -sm gs
-      - shell_command:
-          - python3 passivbot.py binance_01 FTMBUSD /root/apollo/config.json -lm gs -sm gs
-      - shell_command:
-          - python3 passivbot.py binance_01 FTTBUSD /root/apollo/config.json -lm gs -sm gs
-      - shell_command:
-          - python3 passivbot.py binance_01 GALABUSD /root/apollo/config.json -lm gs -sm gs
-      - shell_command:
-          - python3 passivbot.py binance_01 GALBUSD /root/apollo/config.json -lm gs -sm gs
-      - shell_command:
-          - python3 passivbot.py binance_01 GMTBUSD /root/apollo/config.json -lm gs -sm gs
-      - shell_command:
-          - python3 passivbot.py binance_01 ICPBUSD /root/apollo/config.json -lm gs -sm gs
-      - shell_command:
-          - python3 passivbot.py binance_01 LINKBUSD /root/apollo/config.json -lm gs -sm gs
-      - shell_command:
-          - python3 passivbot.py binance_01 LUNA2BUSD /root/apollo/config.json -lm gs -sm gs
-      - shell_command:
-          - python3 passivbot.py binance_01 NEARBUSD /root/apollo/config.json -lm gs -sm gs
-      - shell_command:
-          - python3 passivbot.py binance_01 TLMBUSD /root/apollo/config.json -lm gs -sm gs
-      - shell_command:
-          - python3 passivbot.py binance_01 TRXBUSD /root/apollo/config.json -lm gs -sm gs
-      - shell_command:
-          - python3 passivbot.py binance_01 WAVESBUSD /root/apollo/config.json -lm gs -sm gs
-      - shell_command:
-          - python3 passivbot.py binance_01 XRPBUSD /root/apollo/config.json -lm gs -sm gs
+          - python3 manager start -a
   - window_name: futuresboard
     layout: even-vertical
     shell_command_before:
