+++ sh/git_pull.sh
@@ -3,5 +3,5 @@
 tmux kill-session
 cd /root/apollo
 git stash
-git pull https://github.com/ka1myk/apollo.git onlyRandom
+git pull https://github.com/ka1myk/apollo.git
 chmod +x *.sh
