+++ variables.json
@@ -1,5 +1,1 @@
-{
-  "multiplier": 1.1,
-  "time_to_cool_down": 3600,
-  "exception_cool_down": 10
-}
+{"coins": ["1000LUNCBUSD", "ADABUSD", "ANCBUSD", "APEBUSD", "AVAXBUSD", "BNBBUSD", "DODOBUSD", "DOGEBUSD", "DOTBUSD", "ETHBUSD", "FTMBUSD", "FTTBUSD", "GALABUSD", "GALBUSD", "GMTBUSD", "ICPBUSD", "LINKBUSD", "LUNA2BUSD", "NEARBUSD", "TLMBUSD", "TRXBUSD", "WAVESBUSD", "XRPBUSD"], "multiplier": 1.3, "time_to_cool_down": 13, "exception_cool_down": 5}
